pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toto-interp"
version = "0.1.0"
description = "Independent interpretability research tooling for Datadog Toto"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "toto-ts==0.2.0",
    "datasets==2.17.1",
    "gluonts[torch]==0.16.2",
    "huggingface_hub>=0.31.0",
    "numpy==1.26.4",
    "pandas==2.2.3",
    "scikit-learn==1.5.0",
    "torch==2.7.0",
]

[tool.setuptools.packages.find]
include = ["toto_interp*"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
